Updated README; final cleanup tasks
