+++ ApplicationInsightsRequestLogging/ApplicationInsightsRequestLogging/TelemetryWriter/ITelemetryWriter.cs
@@ -1,22 +1,9 @@
-﻿using Microsoft.ApplicationInsights.DataContracts;
-using Microsoft.AspNetCore.Http;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using Microsoft.AspNetCore.Http;
 
 namespace Azureblue.ApplicationInsights.RequestLogging
 {
     public interface ITelemetryWriter
     {
         public void Write(HttpContext context, string key, string value);
-    }
-
-    public class TelemetryWriter : ITelemetryWriter
-    {
-        public void Write(HttpContext context, string key, string value)
-        {
-            var requestTelemtry = context.Features.Get<RequestTelemetry>();
-            requestTelemtry?.Properties.Add(key, value);
-        }
     }
 }
